simplified class hierarchy.
SVGAttribute was used for classes barely related to SVG attrs, so gave it
a new, proper name -- SVGBasicEntity.  did the same to S2SBlockContainer
(SVGContainerEntity), and subclassed it from SVGBasicEntity to make
relation explicit (in most cases they were used in conjunction).

logic for trafos was even more poorly designed: SVGBlockTrafo was used as
a superclass of SVGTransform to denote relation to basic trafos and make it
easy to check class for convertability to matrix (SVGBlockTrafo.update()).