changed preprocessing.
first of all, preprocessing of actual data was needed only by some of
SVGTrafo* classes, so it was done during init of those classes.

second, it wasn't the preprocessing that had to be done during init of
other SVG-related classes -- it was conversion of SVG string form to inner
format.  common practice in such case is to provide class constructor,
which was implemented as a "classmethod" cls.from_raw_data().

because of this, it is possible to construct classes from directly from
data, which seemingly simplifies lots of logic and changes their nature
from a one-shot throwaway to a reusable for some library, like Inkscape
extension, etc.

also note that it is now required to provide proper data during init of
classes -- there's no default None value.  the same was done to dtype where
it shouldn't have had this feature to begin with.