fixed the way containers are checked for presence of objects.
you expect that "x in y" will check for presence of object "x" in container
"y", but SVGContainerEntity.__contains__() resulted in checking of presence
of object *of certain type*.  now SVGContainerEntity.__contains__() behaves
as expected.  to support previous behavior, new method was added to
SVGContainerEntity -- contains_obj_with_dtype().

previous behavior was handy, but inconsistent with rest of Python.